dl_training/requirements.txt
azure-core==1.9.0
azure-common==1.1.26
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==11.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.3.0
azureml-core==1.19.0
azureml-widgets
click==7.1.2
contextlib2==0.6.0.post1
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
docker==4.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
networkx==2.5
numpy==1.19.4
opt-einsum==3.3.0
pandas==1.1.4
pickleshare==0.7.5
Pillow==8.0.1
PyJWT==1.7.1
PyYaml
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2

